moved non fixtures functions out of the fixtures module
